+++ tests/functional/data/contracts/ethereum/local/sub_reverts.json
@@ -1,1 +1,1 @@
-{"abi":[{"inputs":[{"name":"a","type":"uint256"}],"name":"revertStrings","outputs":[{"name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"}],"ast":{"ast_type":"Module","children":[{"ast_type":"FunctionDef","children":[{"ast_type":"arguments","children":[{"ast_type":"arg","children":[{"ast_type":"Name","children":[],"classification":1,"col_offset":21,"end_col_offset":28,"end_lineno":4,"lineno":4,"src":{"jump_code":"","length":7,"start":49}}],"classification":0,"col_offset":18,"end_col_offset":28,"end_lineno":4,"lineno":4,"src":{"jump_code":"","length":10,"start":46}}],"classification":1,"col_offset":18,"end_col_offset":28,"end_lineno":4,"lineno":4,"src":{"jump_code":"","length":10,"start":46}},{"ast_type":"Assert","children":[{"ast_type":"Compare","children":[{"ast_type":"Name","children":[],"classification":1,"col_offset":11,"end_col_offset":12,"end_lineno":5,"lineno":5,"src":{"jump_code":"","length":1,"start":78}},{"ast_type":"NotEq","children":[],"classification":0,"col_offset":11,"end_col_offset":17,"end_lineno":5,"lineno":5,"src":{"jump_code":"","length":6,"start":78}},{"ast_type":"Int","children":[],"classification":0,"col_offset":16,"end_col_offset":17,"end_lineno":5,"lineno":5,"src":{"jump_code":"","length":1,"start":83}}],"classification":0,"col_offset":11,"end_col_offset":17,"end_lineno":5,"lineno":5,"src":{"jump_code":"","length":6,"start":78}}],"classification":0,"col_offset":4,"end_col_offset":17,"end_lineno":5,"lineno":5,"src":{"jump_code":"","length":13,"start":71}},{"ast_type":"Return","children":[{"ast_type":"NameConstant","children":[],"classification":0,"col_offset":11,"end_col_offset":15,"end_lineno":6,"lineno":6,"src":{"jump_code":"","length":4,"start":113}}],"classification":0,"col_offset":4,"end_col_offset":15,"end_lineno":6,"lineno":6,"src":{"jump_code":"","length":11,"start":106}},{"ast_type":"Name","children":[],"classification":1,"col_offset":1,"end_col_offset":9,"end_lineno":3,"lineno":3,"src":{"jump_code":"","length":8,"start":19}},{"ast_type":"Name","children":[],"classification":1,"col_offset":33,"end_col_offset":37,"end_lineno":4,"lineno":4,"src":{"jump_code":"","length":4,"start":61}}],"classification":1,"col_offset":0,"end_col_offset":15,"end_lineno":6,"lineno":4,"name":"revertStrings","src":{"jump_code":"","length":89,"start":28}}],"classification":0,"col_offset":0,"end_col_offset":15,"end_lineno":6,"lineno":1,"name":"revert_strings_v037.vy","src":{"jump_code":"","length":117}},"contractName":"revert_strings_v037","deploymentBytecode":{"bytecode":"0x61005761000f6000396100576000f36003361161000c5761003f565b60003560e01c346100455763d8046e7d811861003d5760243610610045576004351561004557600160405260206040f35b505b60006000fd5b600080fda165767970657283000307000b"},"dev_messages":{"5":"dev: sub-zero"},"devdoc":{},"pcmap":{"11":{"dev":"dev: Fallback not defined","location":null},"20":{"dev":"dev: USER_ASSERT","location":null},"43":{"location":[5,11,5,12]},"45":{"location":[5,11,5,17]},"46":{"location":[5,11,5,17]},"47":{"location":[5,4,5,17]},"50":{"dev":"dev: USER_ASSERT","location":[5,4,5,17]},"51":{"location":[6,11,6,15]}},"runtimeBytecode":{"bytecode":"0x6003361161000c5761003f565b60003560e01c346100455763d8046e7d811861003d5760243610610045576004351561004557600160405260206040f35b505b60006000fd5b600080fda165767970657283000307000b"},"sourceId":"revert_strings_v037.vy","sourcemap":"-1:-1:0:-;;;;:::-;;:::-;:::-;;;;;;;:::-;;;;;:::-;;;;;:::-;78:1;:6;;71:13;:::-;113:4;-1:-1;;;;;:::-;;:::-;;;;:::-;;;","userdoc":{}}
+{"abi":[{"inputs":[{"name":"a","type":"uint256"}],"name":"revertStrings","outputs":[{"name":"","type":"bool"}],"stateMutability":"nonpayable","type":"function"}],"ast":{"ast_type":"Module","children":[{"ast_type":"FunctionDef","children":[{"ast_type":"arguments","children":[{"ast_type":"arg","children":[{"ast_type":"Name","children":[],"classification":1,"col_offset":21,"end_col_offset":28,"end_lineno":2,"lineno":2,"src":{"contract_id":1,"jump_code":"","length":7,"start":31}}],"classification":0,"col_offset":18,"end_col_offset":28,"end_lineno":2,"lineno":2,"src":{"contract_id":1,"jump_code":"","length":10,"start":28}}],"classification":1,"col_offset":18,"end_col_offset":28,"end_lineno":2,"lineno":2,"src":{"contract_id":1,"jump_code":"","length":10,"start":28}},{"ast_type":"Assert","children":[{"ast_type":"Compare","children":[{"ast_type":"Name","children":[],"classification":1,"col_offset":11,"end_col_offset":12,"end_lineno":3,"lineno":3,"src":{"contract_id":1,"jump_code":"","length":1,"start":60}},{"ast_type":"NotEq","children":[],"classification":0,"col_offset":11,"end_col_offset":17,"end_lineno":3,"lineno":3,"src":{"contract_id":1,"jump_code":"","length":6,"start":60}},{"ast_type":"Int","children":[],"classification":0,"col_offset":16,"end_col_offset":17,"end_lineno":3,"lineno":3,"src":{"contract_id":1,"jump_code":"","length":1,"start":65}}],"classification":0,"col_offset":11,"end_col_offset":17,"end_lineno":3,"lineno":3,"src":{"contract_id":1,"jump_code":"","length":6,"start":60}}],"classification":0,"col_offset":4,"end_col_offset":17,"end_lineno":3,"lineno":3,"src":{"contract_id":1,"jump_code":"","length":13,"start":53}},{"ast_type":"Return","children":[{"ast_type":"NameConstant","children":[],"classification":0,"col_offset":11,"end_col_offset":15,"end_lineno":4,"lineno":4,"src":{"contract_id":1,"jump_code":"","length":4,"start":95}}],"classification":0,"col_offset":4,"end_col_offset":15,"end_lineno":4,"lineno":4,"src":{"contract_id":1,"jump_code":"","length":11,"start":88}},{"ast_type":"Name","children":[],"classification":1,"col_offset":1,"end_col_offset":9,"end_lineno":1,"lineno":1,"src":{"contract_id":1,"jump_code":"","length":8,"start":1}},{"ast_type":"Name","children":[],"classification":1,"col_offset":33,"end_col_offset":37,"end_lineno":2,"lineno":2,"src":{"contract_id":1,"jump_code":"","length":4,"start":43}}],"classification":1,"col_offset":0,"end_col_offset":15,"end_lineno":4,"lineno":2,"name":"revertStrings","src":{"contract_id":1,"jump_code":"","length":89,"start":10}}],"classification":0,"col_offset":0,"end_col_offset":15,"end_lineno":4,"lineno":1,"name":"SubRevertsVy.vy","src":{"contract_id":1,"jump_code":"","length":99}},"contractName":"SubRevertsVy","deploymentBytecode":{"bytecode":"0x61005761000f6000396100576000f36003361161000c5761003f565b60003560e01c346100455763d8046e7d811861003d5760243610610045576004351561004557600160405260206040f35b505b60006000fd5b600080fda165767970657283000307000b"},"dev_messages":{"3":"dev: sub-zero"},"devdoc":{},"pcmap":{"11":{"dev":"dev: Fallback not defined","location":null},"20":{"dev":"dev: USER_ASSERT","location":null},"43":{"location":[3,11,3,12]},"45":{"location":[3,11,3,17]},"46":{"location":[3,11,3,17]},"47":{"location":[3,4,3,17]},"50":{"dev":"dev: USER_ASSERT","location":[3,4,3,17]},"51":{"dev":"dev: USER_ASSERT","location":[4,11,4,15]}},"runtimeBytecode":{"bytecode":"0x6003361161000c5761003f565b60003560e01c346100455763d8046e7d811861003d5760243610610045576004351561004557600160405260206040f35b505b60006000fd5b600080fda165767970657283000307000b"},"sourceId":"SubRevertsVy.vy","sourcemap":"-1:-1:1:-;;;;:::-;;:::-;:::-;;;;;;;:::-;;;;;:::-;;;;;:::-;60:1;:6;;53:13;:::-;95:4;-1:-1;;;;;:::-;;:::-;;;;:::-;;;","userdoc":{}}
